refactor: [#44] extract gpg command wrappers
