+++ src/nautilus_librarian/mods/gpg/domain/exceptions.py
@@ -1,0 +1,10 @@
+class InvalidFingerprint(ValueError):
+    pass
+
+
+class InvalidKeygrip(ValueError):
+    pass
+
+
+class DirNotFound(ValueError):
+    pass
